+++ src/main/java/org/organicdesign/fp/collections/UnmodList.java
@@ -89,10 +89,18 @@ public interface UnmodList<E> extends List<E>, UnmodSortedCollection<E> {
     }
 
     /**
-     The default implementation of this method has O(this.size()) performance.
+     This method is deprecated because implementing it on a List has O(n) performance.  This method
+     will never go away because it's declared on java.util.Collection which List extends.  It's
+     deprecated because you should never use it.
 
+     If you need repeated or fast contains() tests, use a Set instead instead of a List.
+     SortedSet.contains() has O(log2 n) performance.  HashSet.contains() has O(1) performance!
+     If you truly need a one-shot contains test, iterate the list manually, or override the
+     deprecation warning, but include a description of why you need to use a List instead of some
+     kind of Set or Map!
      {@inheritDoc}
      */
+    @Deprecated
     @Override default boolean contains(Object o) {
         for (Object item : this) {
             if (Objects.equals(item, o)) { return true; }