+++ changeLog.md
@@ -1,6 +1,7 @@
 **2015-11-07 0.10.19**: There are many changes in this point release, but unless you are writing your own collections by subclassing the Unmod interfaces, you probably won't notice.
 The main push at this point is near 100% test coverage before a 1.0 release.  NOTE: this has not been released to Maven Central yet.
 
+ - Deprecated List.contains().  See note there.
  - Made null hash to 0 instead of Integer.MIN_VALUE in Equator and ComparisonContext.
  - Changed ComparisonContext helper method names from le to lte (for less-than-or-equal-to) and ge to gte.
  - Allowed ImSortedMap.entrySet() to return and UnmodSortedSet instead of an ImSortedSet.  I may go back on this.